feat: B1M2-3279 add betadda theme
